Cut out a bit more boiler plate for ease of testing
