+++ lib/patronus_fati/data_models/common_state.rb
@@ -1,6 +1,24 @@
 module PatronusFati
   module DataModels
     module CommonState
+      module KlassMethods
+        def [](key)
+          instances[key] ||= new(key)
+        end
+
+        def exists?(mac)
+          instances.key?(mac)
+        end
+
+        def instances
+          @instances ||= {}
+        end
+      end
+
+      def self.included(klass)
+        klass.extend(KlassMethods)
+      end
+
       def active?
         presence.visible_since?(self.class.current_expiration_threshold)
       end
@@ -9,8 +27,21 @@ module PatronusFati
         sync_flag?(:dirtyAttributes) || sync_flag?(:dirtyChildren)
       end
 
+      def diagnostic_data
+        {
+          sync_status: sync_status,
+          current_presence: presence.current_presence.bits,
+          last_presence: presence.last_presence.bits
+        }
+      end
+
       def dirty?
         new? || data_dirty? || status_dirty?
+      end
+
+      def initialize(*_args)
+        self.presence = Presence.new
+        self.sync_status = 0
       end
 
       def mark_synced
